[Ionut] - Multithreaded bounding box calculations for skinned submeshes - Better PhysicsActor <-> PhysicsComponent coupling - Changed the ThreadPool model to a priority based system - Added a ScopedTimer class to simplify timing of functions - Added a reset system to WarScene to unload and reload all units and AIEntities
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@427 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 06fd605090a1d980fa8cd11362c0ec30983f279a [formerly 86cd87aa1a4430bdf775499b8133010bb3360793] [formerly b71b6a2dac6fdfc231dd2a13c889d9ec31468456 [formerly c9f2c3419fd9ef48feec84b8de5598a1a8619f50]]
Former-commit-id: 9b76c4ca580506d2f906c918a13452fdfb4c1586 [formerly 2735020bd8c4c2cff147ca20a9c7a297958424ed]
Former-commit-id: fb6ab872e79492c1487fbf8ada99e4607385fe82

+++ Source Code/Scenes/WarScene/WarSceneAI.cpp
@@ -13,6 +13,10 @@ namespace {
     static std::atomic_bool g_navMeshStarted;
 };
 
+void WarScene::registerPoint(U8 teamID) {
+    _resetUnits = true;
+}
+
 bool WarScene::initializeAI(bool continueOnErrors) {
     //------------------------Artificial Intelligence------------------------//
     // Create 2 AI teams
@@ -23,6 +27,50 @@ bool WarScene::initializeAI(bool continueOnErrors) {
     _faction[0]->addEnemyTeam(_faction[1]->getTeamID());
     _faction[1]->addEnemyTeam(_faction[0]->getTeamID());
 
+    bool state = addUnits();
+    if (state || continueOnErrors) {
+        Scene::initializeAI(continueOnErrors);
+    }
+
+    SceneGraphNode* lightNode = _sceneGraph.findNode("Soldier1");
+    SceneGraphNode* animalNode = _sceneGraph.findNode("Soldier2");
+    SceneGraphNode* heavyNode = _sceneGraph.findNode("Soldier3");
+    lightNode->setActive(false);
+    animalNode->setActive(false);
+    heavyNode->setActive(false);
+
+    return state;
+}
+
+bool WarScene::removeUnits(bool removeNodesOnCall) {
+    WAIT_FOR_CONDITION(!AI::AIManager::getInstance().updating());
+
+    for (U8 i = 0; i < 2; ++i) {
+        if (removeNodesOnCall) {
+            for (NPC* npc : _armyNPCs[i]) {
+                SceneGraphNode* node = npc->getBoundNode();
+                _sceneGraph.getRoot().deleteNode(node);
+            }
+        } else {
+            for (NPC* npc : _armyNPCs[i]) {
+                _sceneGraph.addToDeletionQueue(npc->getBoundNode());
+            }
+        }
+    }
+
+    for (U8 i = 0; i < 2; ++i) {
+        for (AI::AIEntity* const entity : _army[i]) {
+            AI::AIManager::getInstance().unregisterEntity(entity);
+        }
+
+        MemoryManager::DELETE_VECTOR(_army[i]);
+        MemoryManager::DELETE_VECTOR(_armyNPCs[i]);
+    }
+
+    return true;
+}
+
+bool WarScene::addUnits() {
     SceneGraphNode* lightNode = _sceneGraph.findNode("Soldier1");
     SceneGraphNode* animalNode = _sceneGraph.findNode("Soldier2");
     SceneGraphNode* heavyNode = _sceneGraph.findNode("Soldier3");
@@ -220,7 +268,8 @@ bool WarScene::initializeAI(bool continueOnErrors) {
             soldier->setAttribute(to_uint(AI::UnitAttributes::HEALTH_POINTS), 100);
             soldier->setAttribute(to_uint(AI::UnitAttributes::DAMAGE), damage);
             soldier->setAttribute(to_uint(AI::UnitAttributes::ALIVE_FLAG), 1);
-            
+            soldier->setMovementSpeed(speed);
+
             _army[k].push_back(aiSoldier);
             _armyNPCs[k].push_back(soldier);
             
@@ -233,36 +282,30 @@ bool WarScene::initializeAI(bool continueOnErrors) {
             AI::AIManager::getInstance().registerEntity(i, _army[i][j]);
         }
     }
-    bool state = !(_army[0].empty() || _army[1].empty());
 
-    if (state || continueOnErrors) {
-        Scene::initializeAI(continueOnErrors);
+    return !(_army[0].empty() || _army[1].empty());
+}
+
+bool WarScene::resetUnits() {
+    AI::AIManager::getInstance().pauseUpdate(true);
+    bool state = false;
+    if (removeUnits(true)) {
+       state = addUnits();
     }
-
-    _sceneGraph.getRoot().deleteNode(lightNode);
-    _sceneGraph.getRoot().deleteNode(animalNode);
-    _sceneGraph.getRoot().deleteNode(heavyNode);
-
+    AI::AIManager::getInstance().pauseUpdate(false);
     return state;
 }
 
 bool WarScene::deinitializeAI(bool continueOnErrors) {
     AI::AIManager::getInstance().pauseUpdate(true);
-
-    while (AI::AIManager::getInstance().updating()) {
-    }
-
-    for (U8 i = 0; i < 2; ++i) {
-        for (AI::AIEntity* const entity : _army[i]) {
-            AI::AIManager::getInstance().unregisterEntity(entity);
+    if (removeUnits(true)) {
+        for (U8 i = 0; i < 2; ++i) {
+            MemoryManager::DELETE(_faction[i]);
         }
-
-        MemoryManager::DELETE_VECTOR(_army[i]);
-        MemoryManager::DELETE_VECTOR(_armyNPCs[i]);
-        MemoryManager::DELETE(_faction[i]);
+        return Scene::deinitializeAI(continueOnErrors);
     }
 
-    return Scene::deinitializeAI(continueOnErrors);
+    return false;
 }
 
 void WarScene::startSimulation() {
@@ -350,6 +393,7 @@ void WarScene::startSimulation() {
         _infoBox->setMessageType(GUIMessageBox::MessageType::MESSAGE_WARNING);
         _infoBox->show();
     }
+
     AI::AIManager::getInstance().pauseUpdate(false);
 }
